+++ onem2mlib/mcarequests.py
@@ -194,7 +194,7 @@ def _getPath(session, path):
 
 def _isValidResource(resource):
 	return	(resource.type == CON.Type_CSEBase and resource.session) or \
-			(resource.session is not None and resource.session.connected is not None and ( \
+			(resource.session is not None and ( \
 				(resource.parent is not None and resource.parent.resourceID is not None)\
 			) )